# keystore: improve caching mechanism