--- a/internal/api/router.go
+++ b/internal/api/router.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/minio/kes-go"
 	"github.com/minio/kes/internal/auth"
-	"github.com/minio/kes/internal/key"
+	"github.com/minio/kes/internal/keystore"
 	"github.com/minio/kes/internal/log"
 	"github.com/minio/kes/internal/metric"
 	"github.com/minio/kes/internal/sys"
@@ -34,7 +34,7 @@ type RouterConfig struct {
 // EdgeRouterConfig is a structure containing the
 // API configuration for a KES edge server.
 type EdgeRouterConfig struct {
-	Keys *key.Cache
+	Keys *keystore.Cache
 
 	Policies auth.PolicySet